Changed to usize.

+++ src/linux/inputs.rs
@@ -4,7 +4,7 @@ use crate::public::{
 };
 
 impl From<KeybdKey> for u64 {
-    fn from(key: KeybdKey) -> u64 {
+    fn from(key: KeybdKey) -> usize {
         match key {
             BackspaceKey => 0xFF08,
             TabKey => 0xFF09,
